pyproject.toml
[project]
name = "tg-userbot-tester"
version = "0.1.0"
description = "MTProto userbot + proxy bot for Telegram bot interaction"
requires-python = ">=3.10"

dependencies = [
    "telethon>=1.34.0",
    "aiogram>=3.7.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
userbot-tester = "userbot_tester.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100